:recycle: move vec3 to own module
